@@ -3,10 +3,10 @@ import 'package:flutter/foundation.dart';
 
 class ApiRequest {
   final String url;
-  final Map data;
+  final Map? data;
 
   ApiRequest({
-    @required this.url,
+    required this.url,
     this.data,
   });
 
@@ -18,11 +18,11 @@ class ApiRequest {
   }
 
   void get({
-    Function() beforeSend,
-    Function(dynamic data) onSuccess,
-    Function(dynamic error) onError,
+    Function()? beforeSend,
+    Function(dynamic data)? onSuccess,
+    Function(dynamic error)? onError,
   }) {
-    _dio().get(this.url, queryParameters: this.data).then((res) {
+    _dio().get(this.url, queryParameters: this.data as Map<String, dynamic>?).then((res) {
       if (onSuccess != null) onSuccess(res.data);
     }).catchError((error) {
       if (onError != null) onError(error);
