@@ -2,17 +2,17 @@ import 'package:flutter/material.dart';
 
 class LoadingOverlay extends StatelessWidget {
   final Widget child;
-  final bool isLoading;
+  final bool? isLoading;
 
   const LoadingOverlay({
-    Key key,
+    Key? key,
     this.isLoading,
-    @required this.child,
+    required this.child,
   })  : assert(child != null),
         super(key: key);
   @override
   Widget build(BuildContext context) {
-    if (isLoading) {
+    if (isLoading!) {
       return Center(
         child: CircularProgressIndicator(),
       );
