@@ -4,9 +4,9 @@ import 'package:flutter/widgets.dart';
 import 'package:getx_app/models/post.dart';
 
 class PostsListItem extends StatelessWidget {
-  final Post post;
+  final Post? post;
 
-  const PostsListItem({Key key, this.post}) : super(key: key);
+  const PostsListItem({Key? key, this.post}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -31,7 +31,7 @@ class PostsListItem extends StatelessWidget {
             children: [
               ClipOval(
                 child: Image.network(
-                  "https://picsum.photos/50/50?random=${post.id}",
+                  "https://picsum.photos/50/50?random=${post!.id}",
                   height: 40,
                   width: 40,
                 ),
@@ -42,7 +42,7 @@ class PostsListItem extends StatelessWidget {
                   crossAxisAlignment: CrossAxisAlignment.start,
                   children: [
                     Text(
-                      post.title,
+                      post!.title!,
                       overflow: TextOverflow.ellipsis,
                       maxLines: 1,
                       softWrap: false,
@@ -63,7 +63,7 @@ class PostsListItem extends StatelessWidget {
           ),
           SizedBox(height: 15),
           Text(
-            post.body,
+            post!.body!,
             style: TextStyle(color: Colors.black87),
           ),
           Row(
